Fix checkout backend for Cloudflare Workers (not classic Pages)
The connected project is a Worker with static assets, not classic Pages —
Workers doesn't auto-detect a /functions folder the way Pages did. Replaces
functions/api/*.js (Pages Functions convention) with a single src/worker.js
entry point that explicitly routes /api/* via wrangler.jsonc's
run_worker_first, falling through to static asset serving for everything
else. Adds .assetsignore so source/server files (scripts/, src/, supabase/,
package.json, etc.) aren't uploaded as publicly downloadable static files.

README updated to describe the Workers (not Pages) setup flow, including
the branch-selection gotcha that's currently blocking testing.
