+++ package.json
@@ -15,5 +15,8 @@
   },
   "dependencies": {
     "stripe": "^18.0.0"
+  },
+  "devDependencies": {
+    "wrangler": "^4.20.0"
   }
 }
